Lsb 3139 import samples only

+++ PhorcysData/DTOs/ShearwaterDiveSummaryDto.cs
@@ -52,7 +52,7 @@ namespace Phorcys.Data.DTOs
         public int? CnsAfterPercent { get; set; }
 
         /// <summary>Battery level/voltage at end of dive (as integer percent or voltage).</summary>
-        public float? BatteryVoltage { get; set; }
+        public decimal? BatteryVoltage { get; set; }
 
         /// <summary>Time-series profile samples parsed from the CSV (rows 4+). May be empty if the file has no sample data.</summary>
         public List<LogSampleDto> Samples { get; set; } = new List<LogSampleDto>();
